refactor: centralize listener setup in new module
Created Hoard.Listeners module to centralize all event listener
initialization, moving runListeners and runServer helper functions
out of Main.hs.

This simplifies Main.hs by removing inline listener setup code and
provides a cleaner separation of concerns. The runServer function
now properly forks itself to run concurrently.
