+++ hoard.cabal
@@ -38,6 +38,7 @@ library
       Hoard.Events.Collector
       Hoard.Events.HeaderReceived
       Hoard.Events.Node
+      Hoard.Listeners
       Hoard.Listeners.ChainSyncEventListener
       Hoard.Listeners.CollectorEventListener
       Hoard.Listeners.HeaderReceivedListener
